Fix Proper Character relationship with User Entity

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using DotNetCore_RPG.Data;
 using DotNetCore_RPG.Dtos.Character;
 using DotNetCore_RPG.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace DotNetCore_RPG.Services.CharacterService
@@ -18,30 +20,35 @@ namespace DotNetCore_RPG.Services.CharacterService
         };
         private readonly IMapper _mapper;
         private readonly RpgDbContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, RpgDbContext context)
+        public CharacterService(IMapper mapper, RpgDbContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
 
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
 
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             await _context.characters.AddAsync(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.characters.Where(c => c.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.characters.Where(c => c.User.Id == userId).ToListAsync();
+            List<Character> dbCharacters = await _context.characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -49,7 +56,8 @@ namespace DotNetCore_RPG.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.characters.FirstOrDefaultAsync(i => i.Id == id);
+            Character dbCharacter = await _context.characters
+                .FirstOrDefaultAsync(i => i.Id == id && i.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -59,18 +67,27 @@ namespace DotNetCore_RPG.Services.CharacterService
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await _context.characters.FirstOrDefaultAsync(i => i.Id == updateCharacter.Id);
-                character.Name = updateCharacter.Name;
-                character.Class = updateCharacter.Class;
-                character.Defense = updateCharacter.Defense;
-                character.HitPoints = updateCharacter.HitPoints;
-                character.Intelligence = updateCharacter.Intelligence;
-                character.Strength = updateCharacter.Strength;
+                Character character = await _context.characters.Include(c => c.User).FirstOrDefaultAsync(i => i.Id == updateCharacter.Id);
 
-                _context.characters.Update(character);
-                await _context.SaveChangesAsync();
+                if(character.User.Id == GetUserId())
+                {
+                    character.Name = updateCharacter.Name;
+                    character.Class = updateCharacter.Class;
+                    character.Defense = updateCharacter.Defense;
+                    character.HitPoints = updateCharacter.HitPoints;
+                    character.Intelligence = updateCharacter.Intelligence;
+                    character.Strength = updateCharacter.Strength;
 
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    _context.characters.Update(character);
+                    await _context.SaveChangesAsync();
+
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
@@ -86,11 +103,19 @@ namespace DotNetCore_RPG.Services.CharacterService
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.characters.FirstAsync(i => i.Id == id);
-                _context.characters.Remove(character);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _context.characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.characters
+                    .FirstOrDefaultAsync(i => i.Id == id && i.User.Id == GetUserId());
+                if(character != null)
+                {
+                    _context.characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
